+++ sdk/search/arm-search/src/operations/services.ts
@@ -51,7 +51,7 @@ export class ServicesImpl implements Services {
   }
 
   /**
-   * Gets a list of all search services in the given resource group.
+   * Gets a list of all Search services in the given resource group.
    * @param resourceGroupName The name of the resource group within the current subscription. You can
    *                          obtain this value from the Azure Resource Manager API or the portal.
    * @param options The options parameters.
@@ -121,7 +121,7 @@ export class ServicesImpl implements Services {
   }
 
   /**
-   * Gets a list of all search services in the given subscription.
+   * Gets a list of all Search services in the given subscription.
    * @param options The options parameters.
    */
   public listBySubscription(
@@ -339,7 +339,7 @@ export class ServicesImpl implements Services {
   }
 
   /**
-   * Gets a list of all search services in the given resource group.
+   * Gets a list of all Search services in the given resource group.
    * @param resourceGroupName The name of the resource group within the current subscription. You can
    *                          obtain this value from the Azure Resource Manager API or the portal.
    * @param options The options parameters.
@@ -355,7 +355,7 @@ export class ServicesImpl implements Services {
   }
 
   /**
-   * Gets a list of all search services in the given subscription.
+   * Gets a list of all Search services in the given subscription.
    * @param options The options parameters.
    */
   private _listBySubscription(